Removing dialog details from vm
