+++ app/src/main/java/com/example/bookstats/features/bookdetails/tabs/sessions/list/SessionDetailsDiffCallback.kt
